chore: Adapt tango testcase
