--- concert/tests/integration/tango/test_tango.py
@@ -1,113 +1,0 @@
-import unittest
-import time
-import tango
-import numpy
-
-from tango import AttrQuality, AttrWriteType, DispLevel, DevState, DebugIt
-from tango.server import Device, attribute, command, pipe, device_property
-
-class PowerSupply(Device):
-    voltage = attribute(
-        label="Voltage",
-        dtype=float,
-        display_level=DispLevel.OPERATOR,
-        access=AttrWriteType.READ,
-        unit="V",
-        format="8.4f",
-        doc="the power supply voltage",
-    )
-
-    current = attribute(
-        label="Current",
-        dtype=float,
-        display_level=DispLevel.EXPERT,
-        access=AttrWriteType.READ_WRITE,
-        unit="A",
-        format="8.4f",
-        min_value=0.0,
-        max_value=8.5,
-        min_alarm=0.1,
-        max_alarm=8.4,
-        min_warning=0.5,
-        max_warning=8.0,
-        fget="get_current",
-        fset="set_current",
-        doc="the power supply current",
-    )
-
-    noise = attribute(label="Noise", dtype=((int,),), max_dim_x=1024, max_dim_y=1024)
-
-    info = pipe(label="Info")
-
-    host = device_property(dtype=str)
-    port = device_property(dtype=int, default_value=9788)
-
-    def init_device(self):
-        Device.init_device(self)
-        self.__current = 0.0
-        self.set_state(DevState.STANDBY)
-
-    def read_voltage(self):
-        self.info_stream("read_voltage(%s, %d)", self.host, self.port)
-        return 9.99, time.time(), AttrQuality.ATTR_WARNING
-
-    def get_current(self):
-        return self.__current
-
-    def set_current(self, current):
-        # should set the power supply current
-        self.__current = current
-
-    def read_info(self):
-        return "Information", dict(
-            manufacturer="Tango", model="PS2000", version_number=123
-        )
-
-    @DebugIt()
-    def read_noise(self):
-        return numpy.random.random_integers(1000, size=(100, 100))
-
-    @command
-    def TurnOn(self):
-        # turn on the actual power supply here
-        self.set_state(DevState.ON)
-
-    @command
-    def TurnOff(self):
-        # turn off the actual power supply here
-        self.set_state(DevState.OFF)
-
-    @command(
-        dtype_in=float,
-        doc_in="Ramp target current",
-        dtype_out=bool,
-        doc_out="True if ramping went well, False otherwise",
-    )
-    def Ramp(self, target_current):
-        # should do the ramping
-        return True
-
-
-class TestTango(unittest.TestCase):
-
-    def setUp(self) -> None:
-        dev_info = tango.DbDevInfo()
-        dev_info.server = "PowerSupply/test"
-        dev_info._class = "PowerSupply"
-        dev_info.name = "test/power_supply/1"
-        db = tango.Database()
-        db.add_device(dev_info)
-    
-    def test_tango_dummy(self) -> None:
-        PowerSupply.run_server()
-        ps = tango.DeviceProxy("test/power_supply/1")
-        self.assertTrue(ps.state() == DevState.STANDBY)
-        ps.TurnOn()
-        self.assertTrue(ps.state() == DevState.ON)
-        self.assertTrue(ps.Ramp(2.1))
-        ps.TurnOff()
-        self.assertTrue(ps.state() == DevState.OFF)
-
-
-if __name__ == "__main__":
-    unittest.main()
